apply inheritance patterns
